requirements.txt
click
prometheus-client
python-dotenv
python-telegram-bot
requests
typing
